Hand-written CSV listing some of JRuby's dependencies
For #8725

This is hand-written because I cannot find any machine-parsable source
of dependency information for JRuby.

This file is a translation of the upstream JRuby COPYING file.

Most of the `version` columns are listed "unknown" because the versions
are not specified in the JRuby COPYING file.

I have only randomly sampled a few of these entries to verify
correctness. This list is best-effort, and humans make mistakes, so if
you need a perfect, error-free list, this may help you get started in
your license auditing and research process.